--- packages/base/src/features/story/utils/segmentDisplayMode.ts
@@ -1,12 +1,0 @@
-import type { IStorySegmentLayer } from '@jupytergis/schema';
-
-import type { StorySegmentDisplayMode } from '@/src/features/story/types/listStoryScrollDrive';
-
-export function getSegmentDisplayMode(
-  activeSlide: IStorySegmentLayer['parameters'] | undefined,
-): StorySegmentDisplayMode {
-  if (activeSlide?.content?.contentMode === 'markdown') {
-    return 'markdown';
-  }
-  return 'map';
-}
